# spec/api-docs/bolsas-pacientes/openapi-specification.yaml
openapi: 3.0.3
info:
  title: CENATE - Módulo Bolsas de Pacientes API
  description: |
    API REST para la gestión del módulo Bolsas de Pacientes del Sistema de Telemedicina CENATE (EsSalud Perú).

    ## ¿Qué son las Bolsas de Pacientes?

    Una **Bolsa** es un conjunto de pacientes agrupados por criterios específicos (diagnóstico, especialidad, urgencia)
    que requieren atención coordinada a través de CENATE. Coordina atenciones médicas remotas para **4.6M asegurados**
    a través de **414 IPRESS**.

    ## Características principales

    - ✅ Importación masiva desde Excel (v1.11.0 con auto-detección inteligente)
    - ✅ Gestión completa CRUD de solicitudes
    - ✅ Auto-enriquecimiento de datos desde BD (IPRESS, RED, Estados)
    - ✅ Soft delete con auditoría completa
    - ✅ Estados de gestión de citas integrados
    - ✅ Asignación de admisionistas y gestores
    - ✅ Filtrado y búsqueda avanzada

    ## Seguridad

    Todos los endpoints requieren autenticación JWT Bearer token.
    Algunos endpoints adicionales requieren permisos MBAC (Module-Based Access Control).

    ## Base de Datos

    - **Host:** 10.0.89.13:5432
    - **Database:** maestro_cenate
    - **Principales tablas:** dim_solicitud_bolsa, dim_tipos_bolsas, bolsa_107_item, dim_ipress, dim_red

  version: 1.11.0
  contact:
    name: Ing. Styp Canto Rondón
    email: stypcanto@essalud.gob.pe
  license:
    name: EsSalud Perú - Uso Interno
servers:
  - url: http://localhost:8080
...
security:
  - bearerAuth: []
paths:
...
  /api/bolsas/solicitudes:
    get:
      tags:
        - Solicitudes de Bolsa
      summary: Listar todas las solicitudes activas
      description: |
        Obtiene todas las solicitudes de bolsa activas (no eliminadas lógicamente).

        **Datos retornados incluyen:**
        - Información del paciente (DNI, nombre, edad calculada, sexo, teléfono, email)
        - Tipo de bolsa (código y descripción)
        - IPRESS y RED enriquecidos desde BD
        - Estado de gestión de citas actual
        - Fechas de solicitud y actualización
      operationId: listarSolicitudes
      responses:
        '200':
          description: Lista de solicitudes
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SolicitudBolsaDTO'
                  message:
                    type: string
                    example: "Solicitudes obtenidas exitosamente"
                  timestamp:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/bolsas/solicitudes/{id}:
    get:
      tags:
        - Solicitudes de Bolsa
      summary: Obtener una solicitud por ID
      operationId: obtenerSolicitudPorId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID de la solicitud
          example: 1523
      responses:
        '200':
          description: Solicitud encontrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    $ref: '#/components/schemas/SolicitudBolsaDTO'
                  message:
                    type: string
                    example: "Solicitud obtenida exitosamente"
                  timestamp:
                    type: string
                    format: date-time
        '404':
          description: Solicitud no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    delete:
      tags:
        - Solicitudes de Bolsa
      summary: Eliminar una solicitud (soft delete)
      description: |
        Elimina lógicamente una solicitud (establece activo = false).
        La solicitud no se borra físicamente de la BD.
      operationId: eliminarSolicitud
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1523
      responses:
        '200':
          description: Solicitud eliminada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  mensaje:
                    type: string
                    example: "Solicitud eliminada exitosamente"
                  idSolicitud:
                    type: integer
                    format: int64
                    example: 1523
                  timestamp:
                    type: string
                    format: date-time
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /api/bolsas/solicitudes/{id}/asignar:
    patch:
      tags:
        - Solicitudes de Bolsa
      summary: Asignar una gestora a una solicitud
      description: Asigna un usuario con rol de gestora a una solicitud específica
      operationId: asignarGestora
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1523
        - name: idGestora
          in: query
          required: true
          schema:
            type: integer
            format: int64
          description: ID de la usuaria gestora (de tabla usuarios)
          example: 45
      responses:
        '200':
          description: Gestora asignada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  mensaje:
                    type: string
                    example: "Gestora asignada exitosamente"
                  idSolicitud:
                    type: integer
                    format: int64
                    example: 1523
                  idGestora:
                    type: integer
                    format: int64
                    example: 45
                  timestamp:
                    type: string
                    format: date-time
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/bolsas/solicitudes/{id}/estado:
    patch:
      tags:
        - Solicitudes de Bolsa
      summary: Cambiar el estado de una solicitud
      description: |
        Actualiza el estado de gestión de citas de una solicitud.
        Estados disponibles están en tabla dim_estados_gestion_citas.
      operationId: cambiarEstadoSolicitud
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1523
        - name: nuevoEstadoId
          in: query
          required: true
          schema:
            type: integer
            format: int64
          description: ID del nuevo estado (de dim_estados_gestion_citas)
          example: 2
      responses:
        '200':
          description: Estado actualizado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  mensaje:
                    type: string
                    example: "Estado actualizado exitosamente"
                  idSolicitud:
                    type: integer
                    format: int64
                    example: 1523
                  nuevoEstadoId:
                    type: integer
                    format: int64
                    example: 2
                  timestamp:
                    type: string
                    format: date-time
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /api/bolsa107/estadisticas:
    get:
      tags:
        - Bolsa 107
      summary: Obtener estadísticas de la Bolsa 107
      description: |
        Retorna estadísticas generales:
        - Total de pacientes
        - Por derivación (Psicología, Medicina)
        - Por ubicación (Lima, Provincia)
      operationId: obtenerEstadisticasBolsa107
      responses:
        '200':
          description: Estadísticas calculadas
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    type: object
                    properties:
                      total:
                        type: integer
                        example: 350
                      psicologia:
                        type: integer
                        example: 180
                      medicina:
                        type: integer
                        example: 150
                      lima:
                        type: integer
                        example: 220
                      provincia:
                        type: integer
                        example: 130
                  message:
                    type: string
                    example: "Estadísticas calculadas exitosamente"
                  timestamp:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /api/import-excel/cargas:
    get:
      tags:
        - Importación Excel
      summary: Obtener lista de todas las cargas importadas
      description: Retorna historial de todas las importaciones realizadas
      operationId: obtenerListaCargas
      responses:
        '200':
          description: Lista de cargas
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id_carga:
                          type: integer
                          format: int64
                        nombre_archivo:
                          type: string
                        fecha_carga:
                          type: string
                          format: date-time
                        usuario_carga:
                          type: string
                        total_filas:
                          type: integer
                        filas_ok:
                          type: integer
                        filas_error:
                          type: integer
                        id_bolsa:
                          type: integer
                        id_servicio:
                          type: integer
                  message:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/import-excel/pacientes/{idCarga}/datos:
    get:
      tags:
        - Importación Excel
      summary: Obtener datos completos de una carga específica
      description: Retorna items + errores de una carga particular
      operationId: obtenerDatosCarga
      parameters:
        - name: idCarga
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 15
      responses:
        '200':
          description: Datos de la carga
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    type: object
                    properties:
                      carga:
                        type: object
                        properties:
                          id_carga:
                            type: integer
                          nombre_archivo:
                            type: string
                          fecha_carga:
                            type: string
                            format: date-time
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Bolsa107ItemDTO'
                      errores:
                        type: array
                        items:
                          type: object
                          properties:
                            fila:
                              type: integer
                            error:
                              type: string
                      total_items:
                        type: integer
                      total_errores:
                        type: integer
                  message:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/import-excel/cargas/{idCarga}:
    delete:
      tags:
        - Importación Excel
      summary: Eliminar una carga
      description: Elimina físicamente o lógicamente una carga de importación
      operationId: eliminarCarga
      parameters:
        - name: idCarga
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 15
      responses:
        '200':
          description: Carga eliminada
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "Carga eliminada correctamente"
                  timestamp:
                    type: string
                    format: date-time
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtenido del endpoint de autenticación.
        Formato: "Bearer {token}"

        El token debe incluirse en el header Authorization:
        ```
        Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        ```
...